Fixed airspeed sensor for testing, works successfully. Next step: scale the airspeed representation to m/s.

+++ Autopilot/Path Manager/PathManager.c
@@ -517,7 +517,7 @@ void copyGPSData(){
         pmData.positionFix = (char)gpsData.positionFix;
     }
     pmData.batteryLevel = getCurrentPercent();
-    pmData.airspeed = getCurrentAirspeed();
+    pmData.airspeed = (float)getCurrentAirspeed();
     pmData.altitude = getAltitude(); //gpsData.altitude; //want to get altitude regardless of if there is new GPS data
     pmData.checkbyteDMA = generatePMDataDMAChecksum();
 }
